test: update image path
