+++ tests/test_cls_postprocess.py
@@ -1,8 +1,9 @@
 import os
 import sys
-import pytest
-import paddle
+
 import numpy as np
+import paddle
+import pytest
 
 current_dir = os.path.dirname(os.path.abspath(__file__))
 sys.path.append(os.path.abspath(os.path.join(current_dir, "..")))
